# admin/modules/iotserver/node/src/api/swagger/swagger.yaml
swagger: '2.0'
info:
  version: '0.0.1'
  title: Sangoma IoT Server API
# during dev, should point to your local machine
host: 192.168.1.115:3000
basePath: /v1
securityDefinitions:
  basicAuth:
    type: basic
schemes:
  # tip: remove http to make production-grade
  - http
  #- https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /log/level:
    x-swagger-router-controller: iot_server
    parameters:
      - name: sessionid
        in: query
        description: swagger session id
        required: false
        type: string
        default: 'swagger_session'
      - name: token
        in: query
        description: swagger token
        required: false
        type: string
        default: 'iZh8grpZbjQZ2JIFyViUfOoVLcOT7XYqZdu1u34frvfNQrL8ifJAHgpgV'
    get:
      description: Returns the current app logging level
      operationId: getLogLevel
      security: []
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/ArrayOfLocations'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
    post:
      description: Sets the app log level
      operationId: setLogLevel
      security: []
      consumes:
        - application/json
        - application/octet-stream
      parameters:
        - name: data
          in: body
          schema:
            type: object
            properties:
              level:
                type: string
                enum: [FATAL, ERROR, WARN, INFO, DEBUG, TRACE]
                description: the log level to set
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Gateway'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
  /all:
    x-swagger-router-controller: iot_server
    parameters:
      - name: sessionid
        in: query
        description: swagger session id
        required: false
        type: string
        default: 'swagger_session'
      - name: token
        in: query
        description: swagger token
        required: false
        type: string
        default: 'iZh8grpZbjQZ2JIFyViUfOoVLcOT7XYqZdu1u34frvfNQrL8ifJAHgpgV'
    get:
      description: Returns list of available objects
      operationId: getAll
      security: []
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/ArrayOfLocations'
        default:
          description: Error
          schema:
            $ref: '#/definitions/ErrorResponse'
...
definitions:
...
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
...
  ArrayOfLocations:
    type: array
    items:
      type: object
      properties: 
        uuid:
          type: string
          description: location unique identifier
        name:
          type: string
          description: location name
        actions:
          type: string
          description: location allowed actions
        temperature:
          type: string
          description: location temperature
        objtype:
          type: string
          description: the object type
...
  Gateway:
    type: object
    properties: 
      uuid:
        type: string
        description: gateway unique identifier
      name:
        type: string
        description: gateway name
      location:
        type: string
        description: uuid of teh location the gateway belongs to
      actions:
        type: string
        description: gateway allowed actions
      objtype:
        type: string
        description: the object type
...